refactor: restructure project folders and move components
