+++ src/index.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { App } from '@app/App';
-import { Providers } from './app/providers/Providers';
-import './styles/index.css';
+import { Providers } from '@app/providers/Providers';
+import '@app/styles/index.css';
 
 const container = document.getElementById('root') as HTMLElement;
 const root = createRoot(container!);
