+++ src/widgets/Header/HeaderRight.tsx
@@ -1,3 +1,3 @@
-import { UserMenu } from '@/components/UserMenu';
+import { UserMenu } from '@/entities/user/ui/UserMenu';
 
 export const HeaderRight = () => <UserMenu />;
